// src/generated/file-layout.ts
export const fileLayout = {
  "file": "WARMAC.MAC",
  "address": 2723,
  "words": 700,
  "mapLine": 743,
  "fields": {
    "stabuf": {
      "offset": 0,
      "words": 640,
      "line": 676,
      "dimensions": [
        {
          "lower": 0,
          "length": 640
        }
      ]
    },
    "debflg": {
      "offset": 644,
      "words": 1,
      "line": 694,
      "dimensions": []
    },
    "dbuf": {
      "offset": 645,
      "words": 3,
      "line": 696,
      "dimensions": [
        {
          "lower": 0,
          "length": 3
        }
      ]
    },
    "dibuf": {
      "offset": 648,
      "words": 3,
      "line": 697,
      "dimensions": [
        {
          "lower": 0,
          "length": 3
        }
      ]
    },
    "tmp": {
      "offset": 651,
      "words": 6,
      "line": 699,
      "dimensions": [
        {
          "lower": 0,
          "length": 6
        }
      ]
    },
    "intblk": {
      "offset": 657,
      "words": 4,
      "line": 702,
      "dimensions": [
        {
          "lower": 0,
          "length": 4
        }
      ]
    },
    "intadr": {
      "offset": 659,
      "words": 1,
      "line": 704,
      "dimensions": []
    },
    "inttyp": {
      "offset": 660,
      "words": 1,
      "line": 705,
      "dimensions": []
    },
    "intsav": {
      "offset": 661,
      "words": 1,
      "line": 707,
      "dimensions": []
    },
    "intflg": {
      "offset": 662,
      "words": 1,
      "line": 708,
      "dimensions": []
    },
    "trpadr": {
      "offset": 663,
      "words": 1,
      "line": 709,
      "dimensions": []
    },
    "savr": {
      "offset": 664,
      "words": 15,
      "line": 710,
      "dimensions": [
        {
          "lower": 0,
          "length": 15
        }
      ]
    },
    "seed": {
      "offset": 679,
      "words": 1,
      "line": 712,
      "dimensions": []
    },
    "fl.ff": {
      "offset": 680,
      "words": 1,
      "line": 716,
      "dimensions": []
    },
    "fo.fnc": {
      "offset": 681,
      "words": 1,
      "line": 719,
      "dimensions": []
    },
    "fo.ios": {
      "offset": 682,
      "words": 1,
      "line": 720,
      "dimensions": []
    },
    "fo.dev": {
      "offset": 683,
      "words": 1,
      "line": 721,
      "dimensions": []
    },
    "fo.brh": {
      "offset": 684,
      "words": 1,
      "line": 722,
      "dimensions": []
    },
    "fo.nbf": {
      "offset": 685,
      "words": 1,
      "line": 723,
      "dimensions": []
    },
    "fo.leb": {
      "offset": 686,
      "words": 1,
      "line": 724,
      "dimensions": []
    },
    "le.nam": {
      "offset": 687,
      "words": 1,
      "line": 727,
      "dimensions": []
    },
    "le.ext": {
      "offset": 688,
      "words": 1,
      "line": 728,
      "dimensions": []
    },
    "le.prv": {
      "offset": 689,
      "words": 1,
      "line": 729,
      "dimensions": []
    },
    "le.ppn": {
      "offset": 690,
      "words": 1,
      "line": 730,
      "dimensions": []
    },
    "pt.fcn": {
      "offset": 691,
      "words": 1,
      "line": 733,
      "dimensions": []
    },
    "pt.swt": {
      "offset": 692,
      "words": 1,
      "line": 734,
      "dimensions": []
    },
    "pt.ppn": {
      "offset": 693,
      "words": 1,
      "line": 735,
      "dimensions": []
    },
    "pt.sfd": {
      "offset": 694,
      "words": 5,
      "line": 736,
      "dimensions": [
        {
          "lower": 0,
          "length": 5
        }
      ]
    },
    "pt.max": {
      "offset": 699,
      "words": 1,
      "line": 737,
      "dimensions": []
    },
    "foblk": {
      "offset": 681,
      "line": 718,
      "words": 6,
      "dimensions": [
        {
          "lower": 0,
          "length": 6
        }
      ]
    },
    "leblk": {
      "offset": 687,
      "line": 726,
      "words": 4,
      "dimensions": [
        {
          "lower": 0,
          "length": 4
        }
      ]
    },
    "ptblk": {
      "offset": 691,
      "line": 732,
      "words": 9,
      "dimensions": [
        {
          "lower": 0,
          "length": 9
        }
      ]
    }
  }
} as const;